fix(snapshot): widen yy_shift_ofst/yy_reduce_ofst to int32_t; demote unused-label to warning
Two corner-case fixes exposed by exercising Lime on the full
PostgreSQL SQL grammar (3,842 states, 3,584 rules, 145,227-entry
action table -- ~50x larger than any in-tree test grammar).

1. yy_shift_ofst / yy_reduce_ofst widened from int16_t to int32_t.

   The PG grammar produces yy_shift_ofst values up to 145,227 and
   yy_reduce_ofst values up to 141,149, both well above INT16_MAX
   (32,767).  With the old int16_t fields those values silently
   wrapped around, mis-routing every shift / reduce on grammars
   over the threshold.  The smaller test grammars (arithmetic,
   COBOL, etc.) all stayed well below 32k entries so this
   regression was invisible -- exactly the kind of corner the
   user warned about when picking a hard test grammar.

   Memory cost on small grammars is at most 4 * nstate bytes (e.g.
   44 bytes for the arithmetic example) -- trivial.

   Files touched:
   - src/snapshot.h          (struct ParserSnapshot fields)
   - include/snapshot_build.h (LimeParserTables fields)
   - src/snapshot_build.c    (dup_i32 helper + use it)
   - src/snapshot_modify.c   (dup_i32_array clone helper)
   - src/parse_engine.c      (find_shift_action / find_reduce_action locals)
   - src/jit_codegen.c       (locals + remove (int16_t) downcast in bounds check)
   - lime.c                  (snapshot emit declares int32_t arrays)
   - tests/test_*.c          (14 hand-rolled snapshot tests updated)

2. "Label X is never used" demoted from error to warning.

   The diagnostic in translate_code() was incrementing errorcnt
   for any RHS/LHS alias that an action body did not reference --
   including alias declarations on rules that have no action body
   at all (e.g. PG's 469 bare_label_keyword(A) ::= KEYWORD. forms
   that mechanically inherit aliases from the original Bison grammar
   but never reference them).  This caused lime to exit non-zero
   on grammars converted from Bison/Yacc / containing leftover
   alias decls, even when the parser tables generate cleanly.

   Fix: gate on rp->noCode (canonical "user wrote no action body"
   flag, set higher up in the same function), and treat the
   diagnostic as warn-only -- still printed for the developer to
   see, but not counted against errorcnt.  Bison treats the same
   condition as a warning.  The hard errors (duplicate label,
   label-after-overwrite, label-same-as-LHS) remain errors.

Tested: 73 prior tests still pass; new test_pg_grammar (gated by
-Dpg_grammar=true, see follow-up commit) reports max shift_ofst
145,227 / max reduce_ofst 141,149, exercising the int32 path.

+++ include/snapshot_build.h
@@ -52,8 +52,8 @@ typedef struct LimeParserTables {
     uint32_t yy_action_count;
     const uint16_t *yy_lookahead;
     uint32_t yy_lookahead_count;
-    const int16_t *yy_shift_ofst;
-    const int16_t *yy_reduce_ofst;
+    const int32_t *yy_shift_ofst;
+    const int32_t *yy_reduce_ofst;
     const uint16_t *yy_default;
     uint32_t nstate;
